requirements.txt
openai==0.28.0
python-dotenv
google-api-python-client
python-telegram-bot